Fix yarn peer warnings
@testing-library/dom is a proper peer that needed to be added.

React & react-dom in the root should not be necessary, but seem to be the only way to silence a peer warning on yarn install.

Both @babel/plugin-syntax-flow & @babel/plugin-transform-react-jsx peer warnings occur even in a fresh CRA app. Adding the dependencies and transitive peer @babel/core@^7.0.0-0 fixes the warnings and doesn't seem to break anything.